+++ clang-tools-extra/clang-tidy/bugprone/BugproneTidyModule.cpp
@@ -26,7 +26,6 @@
 #include "IncorrectRoundingsCheck.h"
 #include "InfiniteLoopCheck.h"
 #include "IntegerDivisionCheck.h"
-#include "IntegerconstantCheck.h"
 #include "LambdaFunctionNameCheck.h"
 #include "MacroParenthesesCheck.h"
 #include "MacroRepeatedSideEffectsCheck.h"
@@ -105,8 +104,6 @@ public:
         "bugprone-infinite-loop");
     CheckFactories.registerCheck<IntegerDivisionCheck>(
         "bugprone-integer-division");
-    CheckFactories.registerCheck<IntegerconstantCheck>(
-        "bugprone-IntegerConstant");
     CheckFactories.registerCheck<LambdaFunctionNameCheck>(
         "bugprone-lambda-function-name");
     CheckFactories.registerCheck<MacroParenthesesCheck>(